smanal code added to do analysis of output per geomtry unit
There have been a number of small changes mostly to deal with more
anomalies in eqdsk files , but also to update the duct modelling
capability and to enable the codes to interface to the COSSAN UQ
package. As part of the latter, SMITER has been shown to work on Mac
with gfortran.

Eqsdsk issues - reading allows for a version of EQDSK containing field
components. Sometimes the software is fooled into thinking these are
present when they are not. To check for obvious errors of this kind,
geoq now produces plottable dumps of the EQDSK before attempting to
analyse the flux contours for plasma centre and edge locations etc.
Some EQDSK also include values of psi for R=0, which
generate infinite field values in the _geoqvolx.vtk files which
paraview does not handle. The new version checks whether R=0 values are
present and deletes them. geoq also does a simple one point check on
the magnetic field to determine its helicity

Duct modelling capability - not directly relevant to SMITER, except
that the fmesh object has had to be moved into the SMITER subdirectory,
and there are other small changes so that geoq can initialise
calculations on a Cartesian grid for duct modelling
The capability to surround geometry by a cube has been debugged and
tested.

COSSAN UQ - again not very relevant to SMITER - vtktfm has been
modified to split CAD models of the first wall into their constituent
parts using body or surface information passed from CAD.

There have been some improvements to the online documentation suggested
by the UQ student, Dominic Calleja.